services/bot/pyproject.toml
[tool.poetry]
name = "resumeforgebot"
version = "0.1.0"
description = "Resume enhancement bot"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.13"
aiogram = "^3.0.0"
redis = "^5.0.0"
kafka-python = "^2.0.0"
python-dotenv = "^1.0.0"